update categories logic and ui

+++ expense-tracker-rest/src/main/java/com/appi147/expensetracker/service/CategoryService.java
@@ -7,11 +7,13 @@ import com.appi147.expensetracker.exception.ForbiddenException;
 import com.appi147.expensetracker.exception.ResourceNotFoundException;
 import com.appi147.expensetracker.model.request.CategoryCreateRequest;
 import com.appi147.expensetracker.repository.CategoryRepository;
+import com.appi147.expensetracker.repository.SubCategoryRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Set;
 
 @Service
 @RequiredArgsConstructor
@@ -19,10 +21,16 @@ import java.util.List;
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final SubCategoryRepository subCategoryRepository;
 
     public List<Category> getAllCategoriesForCurrentUser() {
         User requester = UserContext.getCurrentUser();
-        return categoryRepository.findAllByCreatedBy_UserId(requester.getUserId());
+        List<Category> categories =  categoryRepository.findAllByCreatedBy_UserId(requester.getUserId());
+        Set<Long> usedCategoryIds = subCategoryRepository.findUsedCategoryIdsByUser(requester.getUserId());
+        for (Category category : categories) {
+            category.setDeletable(!usedCategoryIds.contains(category.getCategoryId()));
+        }
+        return categories;
     }
 
     public Category getCategory(Long categoryId) {
@@ -46,7 +54,15 @@ public class CategoryService {
     }
 
     public void deleteCategory(Long categoryId) {
+        User requester = UserContext.getCurrentUser();
         Category category = getCategoryIfOwnedByCurrentUser(categoryId);
+        boolean isUsed = subCategoryRepository
+                .findUsedCategoryIdsByUser(requester.getUserId())
+                .contains(categoryId);
+        if (isUsed) {
+            log.warn("User [{}] attempted to delete category [{}] used in subcategories", requester.getUserId(), categoryId);
+            throw new ForbiddenException("Cannot delete category: it is used in one or more subcategories.");
+        }
         log.info("Deleting category [{}] by user [{}]", categoryId, category.getCreatedBy().getUserId());
         categoryRepository.delete(category);
     }
